+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.1
+- Change namespace to Wwtg99
+- Use wwtg99/class_loader instead of Loader
+- Add key in IDataMapper update
+- Optimize connection and engine
+
 ## 0.1.0
 - Start project
 - Add MedooPlus and MedooEngine